test(jans-cedarling): move `success_test_json` to own file
